Deprecated config.paths defaults.
It is difficult to have path defaults that make sense in all cases.
Using computed configs we can require the values to be set by
applications when they are used. Defaults are now deprecated and will
print a warning. A future major version will throw errors when values
are not set.