Fix exit flow to include proper exit code.
Passing a callback to the logger to schedule a promise to resolve
was not working (the logger API confusingly doesn't accept a callback,
only the logger transport API does). To wait for the logger to
complete and properly schedule code to run to resolve the promise,
the logger streaming API was used instead.

Additionally, async code cannot run on the `exit` event for the
primary process; this was noticed during this fix and remedied
to run on `beforeExit` instead in the case that nothing is
scheduled on the event loop.