Change License section in README.
Removed the copyright claim/range which was confusing bots and humans as
its scope and application was unclear.

Altered the License description text to clarify BSD-3-Clause usage while
noting that some components may differ and use the non-commercial
license.

Individual projects will be updated to reflect the licensing change.